Beginning the manual merge
Not sure why my system was throwing a fit at ion_auth_lang.php being
gone. I just created a blank file and put it in there, to stop it
complaining.

Yes, some of the views are messed up here. That's because I haven't
gone and fixed the automatic merge just yet; that'lll be fixed soon.